Do not attempt to fetch unknown users from ldap or secondary manager
Do not attempt to fetch a missing user from ldap, because this can
result in lots of ldap queries (although on a properly configured wiki
this *shouldn't* happen).

However, this will also kill the secondary password manager. Since that
feature never really worked for us, we accept this.

TODO: remove secondaryPasswordManager alltogether.

Ticket: #7974