Support different config for referral targets
When the primary LDAP server returns a referral, previously LdapContrib
would try to apply the auth config for the primary server to the target
server. This does not work if the target server requires different
credentials.

To work around this, add a setting in which, depending on the referral
URI, different configuration can be overlaid onto the standard
configuration while talking to the target server.

Additionally, to support different GSSAPI authentication, introduce a
setting to customize the path used for the Kerberos credentials cache
(in case the GSSAPI implementation being used is MIT Kerberos 5).